requirements.txt
pyyaml>=6.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
joblib>=1.3.0
docker>=7.0.0
discord.py>=2.3.0
systemd-python>=235; sys_platform == 'linux'
